vastuchitra-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
stripe==7.11.0
python-dotenv==1.0.1
requests==2.31.0
pillow==10.2.0
boto3==1.34.34
STABLE_DIFFUSION_API_KEY=your_actual_api_key_here 